Allow coercion from String to numbers and booleans automatically
I'm working on something that involves parsing YAML and then doing
parameter substitution on the resulting map. I then take that map and
instantiate it via an `UninstantiatedDescribable`. But it's a giant
hassle to navigate the full map to figure out which cases where I've
substituted a parameter need to be coerced from a `String` to an
`int`, `Float`, `boolean`, etc. Since we already have special-case
handling of `URL` and `Result` here, why not do the same for some
other simple cases?

Also, I incrementalized while I was here.