[JENKINS-53825] Choose from ambiguous class names via a simple heuristic
If there are multiple possible classes matching a name on resolution,
filter for just classes that are either inner classes of the model
class, are in the same package as the model class, or are in
subpackages of the model class's package. If more than one class
passes those filters, we'll still fail.

Also starts reporting all possible clases when there are more than two
of them.