Test with Java 21
Java 21 released Sep 19, 2023. We'd like to announce full support for
Java 21 in early October and would like the most used plugins to be
compiling and testing with Java 21.  The acceptance test harness and
plugin bill of materials tests are already passing with Java 21.  This
is a further step to improve plugin readiness for use with Java 21 and
for development with Java 21.

Drops testing of Java 11 in favor of testing with Java 17 and Java 21.

Continues testing both Windows and Linux